refactor: split filter.py into 4 cohesive sibling modules (iter27)
Move 14 module-level functions + their private constants out of filter.py into
filter_fallback.py / filter_groups.py / filter_pack.py / filter_normalize.py,
re-exported via __all__ so resources.lib.filter.<name> still resolves (release_is_pack
for resolver.py, configure_groups_dialog for router.py, etc.). filter.py NLOC 898->479
(under 500).

Patch-safe: every name a test @patches (parse_title_metadata, _get_filter_settings,
filter_results, telemetry, xbmc, the _SORT_SPECS/_pubdate_sort_key group) stayed in
filter.py; only never-patched helpers moved; new modules don't import filter (no cycle).
PTT parsing unchanged.

Codacy complexity 29 -> 28 (file-nloc>500: 6->5). Verified: 2138/2140 tests pass; the
2 "failures" are pre-existing wall-clock timing-assert flakes (test_poll_until_ready_*,
test_resolve_overlaps_bookmark_cleanup_*) that fail identically at the prior commit
baseline under machine load (~8) and pass with free CPU -- NOT caused by this change
(verified by stashing the split and re-running). pylint 10/10, ruff+black clean, vermin 3.8.
